Move advice only used during elasticsearch tests under elasticsearch module and cleanup some unused utility methods
